# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CareConnect API
  description: >
    Clinic appointment and patient-triage API. Implements the contract
    described in docs/spec.md. This file is the single source of truth: on
    any conflict with another doc, this file wins. All endpoints require
    staff authentication (front-desk or doctor) except login — there is no
    patient-facing surface in v1.


    Conventions: all timestamps are UTC ISO-8601 date-times; `date`
    parameters and `ClockTime` values are interpreted in
    `ClinicSettings.clinicTimezone`. Status mapping — 400 malformed or
    schema-invalid request, 401 unauthenticated, 403 permission denied,
    404 not found, 409 state conflict, 422 policy/business-rule violation,
    429 rate limited, 500 unexpected server error (`INTERNAL_ERROR`, no
    details), 503 AI provider unavailable. Every error body is
    `Error` with a `code` from `ErrorCode`.
  version: 0.2.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /doctors:
    get:
      tags: [Doctors]
      summary: List doctors
      parameters:
        - name: specialtyId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Doctor" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Doctors]
      summary: Create a doctor profile
      description: Front-desk only.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DoctorCreate" }
      responses:
        "201":
          description: Doctor created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Doctor" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /doctors/{doctorId}:
    parameters:
      - $ref: "#/components/parameters/DoctorId"
    get:
      tags: [Doctors]
      summary: Get a doctor
      responses:
        "200":
          description: The doctor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Doctor" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Doctors]
      summary: Update a doctor
      description: >
        Front-desk may update any doctor. A doctor may only update their
        own profile (spec.md §1). Setting `active=false` is rejected with
        `AVAILABILITY_CONFLICTS_WITH_APPOINTMENTS` while non-cancelled
        future appointments exist. A `slotLengthMinutes` change affects
        future slot generation only.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DoctorUpdate" }
      responses:
        "200":
          description: Updated doctor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Doctor" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }

  # --------------------------------------------------------- Availability
  /doctors/{doctorId}/availability:
    parameters:
      - $ref: "#/components/parameters/DoctorId"
    get:
      tags: [Availability]
      summary: List weekly availability rules for a doctor
      responses:
        "200":
          description: Availability rules
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Availability" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Availability]
      summary: Add a weekly availability rule
      description: >
        Front-desk, or the doctor for their own availability. A rule may
        not overlap another rule for the same doctor and day.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AvailabilityCreate" }
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Availability" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /doctors/{doctorId}/availability/{availabilityId}:
    parameters:
      - $ref: "#/components/parameters/DoctorId"
      - $ref: "#/components/parameters/AvailabilityId"
    patch:
      tags: [Availability]
      summary: Update a weekly availability rule
      description: >
        Narrowing or moving a rule is rejected with
        `AVAILABILITY_CONFLICTS_WITH_APPOINTMENTS` if it would orphan
        non-cancelled future appointments.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AvailabilityUpdate" }
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Availability" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Availability]
      summary: Remove a weekly availability rule
      description: >
        Rejected with `AVAILABILITY_CONFLICTS_WITH_APPOINTMENTS` if it
        would orphan non-cancelled future appointments.
      responses:
        "204":
          description: Deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /doctors/{doctorId}/availability-exceptions:
    parameters:
      - $ref: "#/components/parameters/DoctorId"
    get:
      tags: [Availability]
      summary: List one-off availability exceptions
      responses:
        "200":
          description: Exceptions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AvailabilityException" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Availability]
      summary: Add a one-off exception (holiday, extra hours)
      description: >
        An `unavailable` exception with no times blocks the whole day and
        is rejected with `AVAILABILITY_CONFLICTS_WITH_APPOINTMENTS` if
        non-cancelled appointments exist that day. `extra_hours` requires
        both times.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AvailabilityExceptionCreate" }
      responses:
        "201":
          description: Exception created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AvailabilityException" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /doctors/{doctorId}/availability-exceptions/{exceptionId}:
    parameters:
      - $ref: "#/components/parameters/DoctorId"
      - $ref: "#/components/parameters/ExceptionId"
    patch:
      tags: [Availability]
      summary: Update a one-off exception
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AvailabilityExceptionUpdate" }
      responses:
        "200":
          description: Updated exception
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AvailabilityException" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Availability]
      summary: Remove a one-off exception
      responses:
        "204":
          description: Deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # -------------------------------------------------------- Specialties
...
  /clinic-settings:
    get:
      tags: [ClinicSettings]
      summary: Get current clinic policy settings
      responses:
        "200":
          description: Clinic settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClinicSettings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [ClinicSettings]
      summary: Update clinic policy settings
      description: Front-desk only (spec.md §1, §4). Written to the audit log.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClinicSettingsUpdate" }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClinicSettings" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ---------------------------------------------------------------- Audit
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DoctorId:
      name: doctorId
      in: path
      required: true
      schema: { type: string, format: uuid }
    AvailabilityId:
      name: availabilityId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ExceptionId:
      name: exceptionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Malformed or schema-invalid request (code VALIDATION_ERROR)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, or expired credentials, or deactivated user
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted (see spec.md §1)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: >
        State conflict — see the operation description for the specific
        `ErrorCode` (e.g. SLOT_ALREADY_BOOKED, PATIENT_ALREADY_BOOKED,
        PATIENT_ALREADY_EXISTS, DOCTOR_ALREADY_EXISTS, SPECIALTY_ALREADY_EXISTS,
        AVAILABILITY_CONFLICTS_WITH_APPOINTMENTS, AVAILABILITY_OVERLAP, USER_ALREADY_EXISTS, APPOINTMENT_NOT_COMPLETED,
        VISIT_NOTE_NOT_WRITABLE, VISIT_NOTE_LOCKED)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    ErrorCode:
      type: string
      description: >
        The catalog of error codes. Each code has exactly one meaning and is
        used consistently everywhere it is raised.
      enum:
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - INVALID_CREDENTIALS
        - FORBIDDEN
        - NOT_FOUND
        - RATE_LIMITED
        - SLOT_ALREADY_BOOKED
        - PATIENT_ALREADY_BOOKED
        - INVALID_SLOT
        - EMERGENCY_JUSTIFICATION_REQUIRED
        - EMERGENCY_NOT_AUTHORIZED
        - CANCELLATION_WINDOW_CLOSED
        - INVALID_TRANSITION
        - FOLLOW_UP_WINDOW_EXCEEDED
        - APPOINTMENT_NOT_COMPLETED
        - PATIENT_ALREADY_EXISTS
        - USER_ALREADY_EXISTS
        - DOCTOR_ALREADY_EXISTS
        - SPECIALTY_ALREADY_EXISTS
        - AVAILABILITY_OVERLAP
        - AVAILABILITY_CONFLICTS_WITH_APPOINTMENTS
        - VISIT_NOTE_NOT_WRITABLE
        - VISIT_NOTE_LOCKED
        - NO_NOTES_TO_DRAFT
        - AI_SERVICE_UNAVAILABLE
        - INTERNAL_ERROR
    Error:
      type: object
      required: [code, message]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message: { type: string }
    ClockTime:
      type: string
      pattern: '^([01]\d|2[0-3]):[0-5]\d$'
      example: "09:30"
      description: Wall-clock time in the clinic time zone (HH:MM).

    # OpenAPI 3.0: `nullable` only works next to `type`, so nullable references need their own
    # schema (a `nullable: true` sibling of `allOf: [$ref]` does not accept null).
    NullableClockTime:
      type: string
      nullable: true
      pattern: '^([01]\d|2[0-3]):[0-5]\d$'
      example: "09:30"
      description: A ClockTime, or null.
...
    Doctor:
      type: object
      required: [id, userId, name, specialtyId, slotLengthMinutes, active]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        name: { type: string }
        specialtyId: { type: string, format: uuid }
        slotLengthMinutes:
          type: integer
          minimum: 5
          description: >
            Doctor-specific slot length; seeded from the specialty default
            but independently editable (spec.md §4). Changes affect future
            slot generation only.
        active: { type: boolean }
    DoctorCreate:
      type: object
      required: [userId, name, specialtyId]
      properties:
        userId: { type: string, format: uuid }
        name: { type: string }
        specialtyId: { type: string, format: uuid }
        slotLengthMinutes:
          type: integer
          minimum: 5
          description: Defaults to the specialty's defaultSlotLengthMinutes if omitted.
    DoctorUpdate:
      type: object
      properties:
        name: { type: string }
        specialtyId: { type: string, format: uuid }
        slotLengthMinutes: { type: integer, minimum: 5 }
        active: { type: boolean }
    DayOfWeek:
      type: string
      enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
    Availability:
      type: object
      required: [id, doctorId, dayOfWeek, startTime, endTime]
      properties:
        id: { type: string, format: uuid }
        doctorId: { type: string, format: uuid }
        dayOfWeek: { $ref: "#/components/schemas/DayOfWeek" }
        startTime: { $ref: "#/components/schemas/ClockTime" }
        endTime:
          allOf: [{ $ref: "#/components/schemas/ClockTime" }]
          description: Must be later than startTime (no overnight rules).
    AvailabilityCreate:
      type: object
      required: [dayOfWeek, startTime, endTime]
      properties:
        dayOfWeek: { $ref: "#/components/schemas/DayOfWeek" }
        startTime: { $ref: "#/components/schemas/ClockTime" }
        endTime: { $ref: "#/components/schemas/ClockTime" }
    AvailabilityUpdate:
      type: object
      properties:
        dayOfWeek: { $ref: "#/components/schemas/DayOfWeek" }
        startTime: { $ref: "#/components/schemas/ClockTime" }
        endTime: { $ref: "#/components/schemas/ClockTime" }
    AvailabilityExceptionType:
      type: string
      enum: [unavailable, extra_hours]
    AvailabilityException:
      type: object
      required: [id, doctorId, date, type]
      properties:
        id: { type: string, format: uuid }
        doctorId: { type: string, format: uuid }
        date: { type: string, format: date }
        type: { $ref: "#/components/schemas/AvailabilityExceptionType" }
        startTime:
          $ref: "#/components/schemas/NullableClockTime"
        endTime:
          $ref: "#/components/schemas/NullableClockTime"
    AvailabilityExceptionCreate:
      type: object
      required: [date, type]
      properties:
        date: { type: string, format: date }
        type: { $ref: "#/components/schemas/AvailabilityExceptionType" }
        startTime:
          $ref: "#/components/schemas/NullableClockTime"
        endTime:
          $ref: "#/components/schemas/NullableClockTime"
    AvailabilityExceptionUpdate:
      type: object
      properties:
        date: { type: string, format: date }
        type: { $ref: "#/components/schemas/AvailabilityExceptionType" }
        startTime:
          $ref: "#/components/schemas/NullableClockTime"
        endTime:
          $ref: "#/components/schemas/NullableClockTime"
...
    ClinicSettings:
      type: object
      required:
        [cancellationCutoffHours, emergencySlotsPerDoctorPerDay, followUpMaxDays, clinicTimezone]
      properties:
        cancellationCutoffHours:
          type: number
          minimum: 0
          default: 2
          description: Governs both cancellation and reschedule (spec.md §4).
        emergencySlotsPerDoctorPerDay:
          type: integer
          minimum: 0
          default: 1
        followUpMaxDays:
          type: integer
          minimum: 1
          default: 30
        clinicTimezone:
          type: string
          example: Asia/Kolkata
          description: IANA time zone name; defines "today" and availability wall-times.
        defaultTriageSpecialtyId:
          type: string
          format: uuid
          nullable: true
          description: >
            Specialty attached to a red-flag triage result returned while
            the AI provider is unavailable (spec.md §5.1).
    ClinicSettingsUpdate:
      type: object
      properties:
        cancellationCutoffHours: { type: number, minimum: 0 }
        emergencySlotsPerDoctorPerDay: { type: integer, minimum: 0 }
        followUpMaxDays: { type: integer, minimum: 1 }
        clinicTimezone: { type: string }
        defaultTriageSpecialtyId: { type: string, format: uuid, nullable: true }
...